core/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
gtts
playsound
pyaudio
pytorch
speechrecognition
torchaudio
torchvision
transformers
webrtcvad